// server/@types/shared/models/OffenderFullDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OffenderDto } from './OffenderDto';
export type OffenderFullDto = (OffenderDto & {
    forename?: string;
    surname?: string;
    middleNames?: Array<string>;
} & {
    forename: string;
    surname: string;
    middleNames: Array<string>;
});