// server/@types/shared/models/ProjectAllocationsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProjectAllocationDto } from './ProjectAllocationDto';
export type ProjectAllocationsDto = {
    /**
     * List of project allocations
     */
    allocations: Array<ProjectAllocationDto>;
};